requirements.txt
absl-py==1.2.0
backcall==0.2.0
cachetools==4.2.4
certifi==2021.5.30
charset-normalizer==2.0.12
click==7.1.2
cycler==0.11.0
Cython==0.29.32
cython-bbox==0.1.3
dataclasses==0.8
decorator==4.4.2
et-xmlfile==1.1.0
filterpy==1.4.5
google-auth==2.11.0
google-auth-oauthlib==0.4.6
grpcio==1.48.1
idna==3.4
imageio==2.15.0
imageio-ffmpeg==0.4.7
importlib-metadata==4.8.3
importlib-resources==5.4.0
ipython==7.16.3
ipython-genutils==0.2.0
jedi==0.17.2
joblib==1.1.0
kiwisolver==1.3.1
lap==0.4.0
Markdown==3.3.7
matplotlib==3.3.4
motmetrics==1.2.5
moviepy==1.0.3
networkx==2.5.1
numpy==1.19.5
oauthlib==3.2.1
opencv-python-headless==4.6.0.66
openpyxl==3.0.10
pandas==1.1.5
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
proglog==0.1.10
prompt-toolkit==3.0.31
protobuf==3.19.5
psutil==5.9.2
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.5
Pygments==2.13.0
pyparsing
PyQt5==5.15.4
pyqt5-plugins==5.15.4.2.2
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.1
pyqt5-tools==5.15.4.3.2
PySide2==5.15.2.1
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.2.1
PyWavelets==1.1.1
PyYAML==6.0
qt5-applications==5.15.2.2.2
qt5-tools==5.15.2.1.2
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.17.2
scikit-learn==0.24.2
scipy==1.5.4
seaborn==0.11.2
shiboken2==5.15.2.1
six==1.16.0
sklearn==0.0
tensorboard==2.10.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
thop==0.1.1.post2209072238
threadpoolctl==3.1.0
tifffile==2020.9.3
torchvision
tqdm==4.64.1
traitlets==4.3.3
typing_extensions==4.1.1
urllib3==1.26.12
wcwidth==0.2.5
Werkzeug==2.0.3
xmltodict==0.13.0
zipp==3.6.0
